% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/error.R
\docType{package}
\name{macBriain}
\alias{macBriain}
\alias{macBriain-package}
\title{Say error messages aloud, on a Mac}
\description{
This package makes R say error messages aloud.
For fun, it also adds some remarks that are considered funny
by some.
}
\details{
To use it, just attach the package with \code{library()} and
the don't forget to unmute your speaker.

To get rid it, unload the package with \code{unloadNamespace()}.
}

